redo call status
Now we use AT+CLIP and AT+COLP to determine call status, rather than
issuing AT+CLCC. This is way simpler, as +CLIP effectively lets
us get the phone number from RING, and +COLP will tell us when the
other end has picked up.

In addition, handle_resp was completely broken, as it didn't
detect newlines properly, and wouldn't process a whole response,
causing it to fall behind. Now it processes everything that is in
the buffer as soon as it can.